doc-extract-backend/requirements.txt
flask
flask-cors
openai
pdfplumber
pandas
sqlalchemy
python-dotenv
pdf2image
pillow
pytesseract
python-dateutil
